requirements.txt
# ================================
# Med3D-MoE-Seg Requirements
# ================================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers & LLM
transformers>=4.35.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Distributed Training
deepspeed>=0.12.0

# PEFT (LoRA)
peft>=0.7.0

# Medical Image Processing
monai>=1.3.0
nibabel>=5.1.0
SimpleITK>=2.3.0
pydicom>=2.4.0

# Numerical & Scientific
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Utilities
einops>=0.7.0
tqdm>=4.66.0
pyyaml>=6.0.1
h5py>=3.9.0

# Visualization & Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
wandb>=0.16.0

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.10.0
flake8>=6.1.0
isort>=5.12.0

# Optional: Segment Anything 3D
# segment-anything-3d  # Install from source if available

# Optional: CT-CLIP dependencies
# Add CT-CLIP specific requirements here if needed
